cleanup: remove unused useRef import from TestDetail
